requirements.txt
requests==2.26.0
pyTelegramBotAPI==4.0.1
python-dotenv==0.19.1
python-magic